+++ philo/include/exec.h
@@ -6,7 +6,7 @@
 /*   By: ldevoude <ldevoude@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/18 15:17:28 by ldevoude          #+#    #+#             */
-/*   Updated: 2025/08/21 16:15:18 by ldevoude         ###   ########lyon.fr   */
+/*   Updated: 2025/08/29 12:55:39 by ldevoude         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,14 +31,21 @@ typedef enum e_msg_to_print
 // philo_monitor.c
 void	*philo_monitor(void *arg);
 
-// philo_routine.c
+// philo_routine_even.c
 void	*routine_odd(void *arg);
+
+// philo_routine_odd.c
 void	*routine_even(void *arg);
+
+// philo_routine
+int wait_all_threads(t_settings *set, t_philo *philo);
+long	fill_now_variable(long *now);
 
 // philo_utils.c
 void	print_msg_routine(t_philo *philo, size_t cases);
 void	routine_take_fork(t_philo *philo, bool right);
 void	ft_usleep(long usec, t_settings *set);
+
 // void routine_sleep(t_philo *philo, bool *first_iteration);
 
 // philo.c
